NES_Copilot_GUI/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
torch>=1.13.0,<2.1.0
sbi==0.22.0
transformers>=4.18.0
joblib>=1.1.0
pebble>=4.6.0
